fix(schedule a trip): dashboard return trip information on trip schedule (#25)
