+++ fasta-backend/package.json
@@ -8,6 +8,7 @@
   },
   "dependencies": {
     "@googlemaps/google-maps-services-js": "^2.6.0",
+    "axios": "^0.19.2",
     "bcrypt": "^4.0.1",
     "body-parser": "^1.19.0",
     "cookie-parser": "~1.4.4",
